working on displaying props
